修改

+++ Dasher/src/com/dasher/service/MarketService.java
@@ -14,5 +14,5 @@ public interface MarketService {
 	public Market getByMarketName(String name);
 	public List<Market> list(String searchStr,int startRow,int pageSize);
 	public List<Market> menuList();
-	public List<Market> getNearList(float longitude,float latitude,float distance);
+	public List<Market> getNearList(double longitude,double latitude,float distance);
 }
